[DM-19540] Checkpoint logging deployment
So far, this will spin up and index the kubernetes logs.

For some reason, the qserv logs look like they are ingested,
but cannot be found in the search frontend.

Auth is currently disabled, but there is support in open distro
so that will be reenabled at a future date.