[DM-22180] Make a science-platform app just for int.
This probably isn't the best organizational style, but it is at least
one.  This allows me to work on the gke version pulling in more things,
while having -int be behind a bit.  Plus we still have to figure out
how many argo clusters we will have and how that should map back.

These point at values files using HTTP links.  This might be interesting,
because I don't know how / if they will get triggered based on changes,
but since they're in the same repo at least as the application, which is
watched, I'm hoping it will work.