[DM-22376] Vaultsecrets for TAP on stable
So this change to the values yaml file will allow us to use the
new version of the helm chart to deploy on stable.  The new helm
chart uses vault for its secrets, and since we installed the vault
secrets operator there, we can now use this over there.  I have
created the secret using the vault command line utility.