[DM-22376] Take control of stable landing-page
Even though we can't do it with the latest image yet (because we
need to have the firefly URL be correct) we can still run the currently
running image with helm.  So let's add in those values, and set
the image to be the currently running image (MOTD_URL should be
ignored, but we will soon be wanting it).