[DM-22463] Override emptyDir to null
This is a little tricky.  So Helm will merge the values from the
portal.yaml into the values.yaml provided by the chart.  Which is
exactly what it is supposed to do.  But for the values part where
the volumes are supplied, they have different keys.  So this kind
of makes one mess where all the keys are included.  By setting
emptyDir: null this means that it will get rid of that key entirely,
which is what kubernetes wants.