[DM-24595] Remove auth from int Argo CD
Temporarily remove authnz from in front of the Argo CD console on
int.  Port forwarding does not work properly to access the console
for that version of Argo CD, which makes it unsafe to upgrade auth
with auth in front of the UI.  This is behind a VPN and has its own
separate password, so this should be safe as a temporary measure
until the upgrade is complete.