Pin Jupyterhub config back to 0.22.0-appropriate values for int/stable
In 0.23, we no longer need to configure auth and spawner classes, because
there's only one of each (because non-JWT support was dropped).  But
since we don't track "latest" on int and stable, we need to keep their
configs the same.  When they move to 0.23 or later, we can drop their
config overrides from values, and just use the charts one.