Support the DNS solver for cert-manager
Add support for the DNS solver and configure the Tucson teststand
to use it.  It cannot use the HTTP solver because it is behind a
firewall.  Currently, this is an option; in the future, we'll switch
all Science Platform environments to the DNS solver for consistency
and better testing.

Configure sync waves for the cert-manager resources so that the
cluster issuer and default certificate are created after the rest
of cert-manager has been configured and passes health checks.  This
will hopefully avoid failures during initial deploy because the
admission web hooks aren't yet running.

The corresponding secret for the Tucson teststand has already been
created in Vault.