Work around design flaws in Helm templating
Helm templates don't like dashes, so change the parameters for
cert-manager configuration to underscores.  Add another layer of
conditional because Helm can't understand that a boolean test on
a non-existent setting should be false and apparently doesn't have
short-circuiting "and" either.

This makes the cluster-issuer.yaml template temporarily messy by
repeating the else block.  I'm not bothering to fix this for now
because this will go away when we standardize on the DNS solver.