[DM-24542] Update generate_secrets for logging
Okay, so in the secret, I'm going to generate the different users
and their passwords (and associated bcrypt hashes).  This is all
stored in the log vault secret, and read from the chart.  This
generates random passwords for each environment.

Also add a requirements.txt since there was not one.