[DM-24542] Get rid of JWT, fix accounts for new secret
Okay, so first, get rid of the JWT configuration for that backend,
and the associated signing key to verify the tokens.  We're just
going to use a hardcoded account, and it will be behind the
gafaelfawr proxy anyway, so this isn't a big thing.  But we do
probably want some kind of account.