Split cert-manager and cert-issuer
Move the certificate issuer resource into a separate application
and remove it from the cert-manager application to handle the base
and summit use case where cert-manager will be installed and managed
by IT rather than by us, but with a default issuer that doesn't
support our domain name.

Remove the default certificate configuration from nginx-ingress and
instead rely on nublado to define a TLS configuration for its ingress
that will be merged with all the other ingresses on the same host.

Enable cert-manager on the test GKE environments since the sync
issues should now be fixed.