Switch bleed.lsst.codes to named virtual hosts
There appears to be no way to avoid using the default certificate
with nginx and also use the * virtual host with TLS certificates
because nginx can't find the appropriate certificate to use and
falls back on the self-signed one.  Convert all the bleed app
installs to naming the ingress host so that TLS works properly.