Switch to named virtual hosts everywhere
Certificate management by configuring the TLS certificate for
nublado's JupyterHub proxy and letting everyone else inherit it
via TLS merging on a virtual host only works with named virtual
hosts because ingress-nginx won't merge TLS for the * virtual host.

Therefore, change all enviroments to named virtual hosts.

This was already the case for the NCSA, summit, base, and Tucson
test stand environments, so this change is only in the GKE
enviroments.  Also add a parameter to the install.sh installer
script for the hostname so that Argo CD can be configured with the
appropriate named virtual host as well.