[DM-27044] Give up trying to mount NFS inside minikube
So first off, there's DNS problems, but those are fixable.  The
problem stems from the host doing the mounting, so therefore the
host must be able to look up the IP address, but the DNS on the
host aparently isn't controlled by k8s for reasons of security
(or some other reason).

Since there's nothing in these anyway, I'm going to give up on this,
but it's not from lack of trying.  Setting up the fileserver pod
locally it would just time out mounting as well, even when I fixed
the DNS.