[DM-27279] Move stable over to small qserv
First, point stable over to the address of the small qserv.
Next, we need to release a new version of the TAP service to roll
out the proper monkey queries as well.  While we don't theoretically
need to roll this out everywhere (since only stable is running
the monkey) I like to keep the spirit of everything up to date.