[DM-27787] Allow vault address to be configurable
First off, instead of using the same values.yaml for every environment,
modify the application to use the environment name to point at
values-{environment}.yaml instead, like other services.  Then replicate
the values.yaml to those files, and delete values.yaml.  This shouldn't
have an effect on what is rendered (it should still be in sync, I think,
since the resources won't be changed).  But if an environment needs to
point to a different vault address, that can be configured in the values
files.