[DM-28043] Fix up minikube networking
Okay, so first off, this came about by not being able to get to the
gafaelfawr token endpoint, due to the annotations being wrong (no
port), and no way to configure it.

Realizing I'm digging a hole rather than making a pile, I found a
way to configure the nginx-ingress to use the host networking on
80 and 443 (which is actually the VM's host).  This means we can
get rid of port 31337 (picked because NodePorts can only have
ephemeral range).

Using the host network also allows the pod to contact itself,
something it couldn't do previously since it wasn't attached to
the host networking stack.

One downside is that I haven't gotten the admissions webhook
service working yet, and this means that you won't get a warning
if you make an editing goof in your ingress yaml.  But only for
minikube.  I think this can be fixed later, but I'm not sure how,
it has to do with binding that to a nodeport or something.

We will also need to notify CILogon to change the redirect URL
to not redirect to 31337 but instead just 443.