[DM-28050] Make repository URL configurable
Okay, so part of the problem with forking this repo is that the
repo URL is checked into a bunch of the templates.  It's all the
same, so first, let's parameterize that out to a values.yaml file
with the default being what it currently was hardcoded to.

But this allows the install.sh script to override that value while
creating the science-platform app, allowing it to point to whatever
the current repository that it is being run from is (and not the
original lsp-deploy).