[DM-28050] Refactor install.sh for helm3
Refactor the installer to use helm3.

First, we don't need the initial-resources.yaml, since there's no
server side component to helm3.  Yay.  Dump that.

Next, use helm3 for the command lines, and install argocd from the
services directory, rather than this argo-cd-values.yaml that is
deleted.  Because we install it from the services directory, and
the resources are the same, argocd won't sync it and will just be
like "oh it's there!"  So in this, argocd doesn't restart, and
therefore it doesn't hang up during the install.  Argocd app
for argocd could be moved outside of the science-platform in a
future commit.

As a bonus, the hostname is also read out of the services directory's
values.yaml file, which means that we can eliminate that parameter.