[DM-28050] Flatten out revision
Okay, this revision tag has gotten out of control, and having an
n by n deployment just hurts my brain.  Luckily no one is doing this,
and everyone is on HEAD (and I can tell, because that's checked in!).

Currently we are passing in REVISION on the command line (which actually
only sets the revision for the science-platform app, and the subapps
will automatically default in the branch to HEAD, since they aren't
overridden).  So it doesn't really work as expected.

What this does is make one revision tag, and sets all the apps to use
the same branch / commit hash.  This makes it a lot easier to understand.
Get rid of all the individual revisions, and replace it with one that is
default to HEAD at the top level.  This is overridden in the installer
to the current branch that your git is at.

This means that when you run the installer, it will use the installer
for your branch (since you have it checked out), and install the
science-platform app at the current branch (passed in with --revision
to argocd create app).  Then all the subapps are configured via helm
and also point to the same branch.  So if your branch is pushed, you
should have it install from there.

But be warned!  If the branch goes away, you will be unhappy.  But you
will be able to run the installer script from a new branch to make it
better.