Enable cert-manager and ingress-nginx at base and summit
For now, we'll be managing cert-manager and ingress-nginx in the
antu and andes clusters.  Enable them and add the required values
files for ingress-nginx.

Try to pin the IP addresses for base, summit, and tucson-teststand
since we hopefully should be able to request the same IP address as
we're currently using in those clusters.