[DM-28482] Don't use uid, user_namespace
Don't use the singleuser uid in the chart, we'll use pod_uid's
default which is in the nublado2 chart now.

Also, we can use the user_namespace template variable now, which
will automatically know and compute the proper namespace for us,
so that means the templating can be more cookie cutter.