[DM-28839] Add vault_path_prefix
This will give us a way to figure out what the enclave and other
keys are in vault.  We don't currently end up using this yet in the
templates, but that's for the future.  Having these checked in the
yaml file means that I can have the installer read it.