Pin mobu and cachemachine charts and remove tags
Rather than pinning the Docker image tag, pin the version of the
chart for mobu and cachemachine.  The chart now pins the Docker
image.  This works correctly with Renovate's automatic PR
generation: it will now create PRs for new Docker image versions
in the charts repo and then create PRs for new chart versions in
the Phalanx repo.