Update Gafaelfawr configuration
The Gafaelfawr chart has been substantially overhauled and a lot
of values parameters have been renamed, moved, or replaced.  Update
the environment configurations accordingly.

Drop group mappings for scopes that were previously defined but
that we're not using anywhere so that the charts are clearer about
which group mappings are in use.  Add explicit group mappings for
the environments using NCSA groups, since the upstream chart no
longer provides a default (that was very confusing).