Reorganize Phalanx documentation
Move all the operational guides into a new ops subdirectory,
matching the layout of Roundtable.  Move the bootstrapping
documentation under ops as well.  Move the architectural
documentation of TLS configuration into the bootstrapping guide,
since that's where it matters.  Move the TLS configuration details
into a bootstrapping note for Squareone, and replace references to
using Nublado for the TLS configuration.