Update to Gafaelfawr 4.2.0
Stop passing name to the Portal in headers.  Upgrade Gafaelfawr to
4.2.0, which drops support for name in HTTP headers due to Unicode
issues and checks that the user is a member of a valid group.  Add
an error footer to the IDF environments pointing to the GitHub
project for issues (this is also on int and dev so that we can test
it more easily, not because those users should use that channel).