Update configuration for mobu 4.0.0
The next release of mobu will not require specifying all the values
to submit to the options form and will default to a Large image.
It also supports testing the latest weekly instead of the recommended
image, so use that on IDF int to test the latest weekly image before
we promote it.