Run workflows on PRs, always run minikube
Run the workflows on PRs rather than pushes so that they will run
for external submissions.  Redo how minikube is run using an action
to filter on changed paths so that the job will always run (and thus
can be a blocking job for merges), but will quickly do nothing if the
paths changed do not affect minikube.