Change application to service
We decided to standardize on the service terminology.  Make that
change to the rest of the documentation, except for places where
we're specifically talking about Argo CD applications.  Add an
explanation to the introduction about the service terminology, the
Argo CD application terminology, and the difference between that
and a Kubernetes Service resource.