Allow updates of secrets from 1Password
This changes the installer/generate_secrets.py script to allow updates of
secrets that already exist in Vault, and have a value in 1Password. Previously,
the value in Vault was always preserved.

There is some special behavior for ingress-nginx and cert-manager which I don't
understand; I'm preserving the old code's behavior (never update those two)
just out of caution.