Update moneypenny configuration
Use the pull-secret chart to install pull-secret instead of a
separate VaultSecret that came with the old moneypenny chart.
Update values-*.yaml for the new configuration options and remove
boilerplate that's now handled by the chart.  Remove the retire
configuration, since we currently don't use it anywhere.

Enable the network policy on the IDF.