Move Gafaelfawr chart into Phalanx
Following RFC-830, move the gafaelfawr chart from the separate
charts repository to Phalanx.  Use the same README.md template
that sasquatch is using.

Drop a bunch of configuration options that aren't necessary now
that the scope of the chart is much smaller and were never used.
Directly create the pull-secret rather than including a chart.