Move redis dependency to top-level chart
Helm was not showing the redis deployment when it was a sub-sub-chart
(via the time-square subchart). Moving redis up to the top-level as a
normal sub-chart should resolve this problem.

It does mean that the redis URL now needs to be hard-coded, but this
shouldn't be a difficulty since the redis deployment is fully under this
Helm chart's control.