Use Helm globals set by Application resource
Use the environment's configurations for host, baseUrl, and
vaultSecretsPathPrefix.

Note that the host is still backed into environment values for TLS; not
sure of the best way to move this up to the environment Argo CD
application level.

Also the pull secret still uses a backed in vault secrets path; this
will get dropped soon when we move Squareone to the GitHub Container
Registry.