Update cert-manager cluster issuer and docs
Remove the cert-issuer docs and merge them with the cert-manager
docs now that the services have been merged.  Rename the cluster
issuer to avoid both the cert-manager and cert-issuer prefix and
just call it letsencrypt-issuer.  Update the installer to stop
trying to sync cert-issuer.  Remove cert-issuer from the
science-platform values files.