Inject butlerRepositoryIndex
Both vo-cutouts and datalinker need DAF_BUTLER_REPOSITORY_INDEX to
resolve Butler locations.  This is a per-site configuration that
should also eventually be injected into nublado2, so add it to the
deployment values file and inject it via Argo CD.  Adjust datalinker
and vo-cutouts charts accordingly, and set the environment variable
on the relevant deployments.

For vo-cutouts, we can now drop CUTOUT_BUTLER_REPOSITORY, since this
replaces that setting.