Remove auth-signin config for hips
I was assuming that this would be accessed via the browser, but
it looks like we're going to use the Portal as the interface and
won't publish an index.html file, making it hard to use from the
browser.  Therefore, don't do the automatic redirect when users
aren't logged in, and instead return a 403.