Add Gafaelfawr maintenance CronJob
This is a bit complicated because the Google Cloud SQL sidecar
normally keeps running and thus keeps the Job from terminating.
Use a hack from StackOverflow to have the main container signal the
sidecar container that it should exit.  That in turn requires using
the Alpine version of the Google sidecar image so that /bin/sh
exists.

Also fix a bug in adding pod annotations to the Kubernetes operator
pod, noticed when copying configuration from it to the new CronJob.