Use separate Cloud SQL Proxy pod
Using a Cloud SQL Proxy sidecar for CronJob resources is very
annoying, since one has to coordinate stopping the Cloud SQL Proxy
server so that the CronJob can exit, which in turn requires a bunch
of opaque shell gunk.

Since we're about to add another CronJob, add a separate Cloud SQL
Proxy service that can be used by all the non-critical, low-volume
services (the Kubernetes operator and the CronJobs to start).
Continue to use the sidecar for the main Gafaelfawr pods so that
the proxy will scale with the pods and the critical Gafaelfawr
service doesn't have a single point of failure.

Add a NetworkPolicy to keep anything else in the cluster from talking
to the Cloud SQL Proxy pod.  This will only work at Google, but this
service is only used at Google, so that should be sufficient.

This requires generating two versions of the Gafaelfawr ConfigMap,
since the database URL has to be different between the Gafaelfawr
pods (using localhost) and the other pods (using the service in the
same namespace).  Do that with multiple resources and a Helm define
inside that template.

Drop the database URL from values.yaml for environments where the
Cloud SQL Proxy is used, since we always know what it is (and have
to adjust it for the two different deployment patterns).