Improve Argo CD application configuration
Inject the Vault secrets path like we do for other applications,
and modify the installer to also inject it (and correctly pass in
both values.yaml files; I'm not sure how it was working before).
Install the Vault secret unconditionally, since it was enabled for
every environment, and remove the configuration from the individual
values files.