Add read:image scopes to T&S deployments
I didn't add these when the scope was added for the cutout service,
since it wasn't clear we would have images in the same sense on T&S
deployments, but now the Portal requests this scope, which means users
have to have it when using the Portal to do image display.  It's
simpler to just grant the scope to the same groups as have exec:portal
access than try to change the scope requirements in those environments.
This matches what we did with read:tap, which also isn't directly in
use there.