Clean up unused injected variables
Remove Argo CD injected Helm variables that aren't used, and
document the ones that are used.  Rename the vault_certificate
setting for ingress-nginx to vaultCertificate, following the naming
rules for Helm chart parameters.

This leaves some unused injected variables for semaphore on the
grounds that they may be used in the future.