Stop specially creating user groups in template
Gafaelfawr now includes the user private group in the user's normal
gorup list, and some deployments, such as USDF, will not use a user
private group with the same GID as the user's UID.  Stop generating
a separate /etc/group entry with a GID equal to the UID and rely on
the groups from Gafaelfawr.

Also change the user's GID to their GID attribute if it is set.