Drop X-Auth-Request-Uid from TAP
Make TAP stop requesting X-Auth-Request-Uid from Gafaelfawr, since
the next release of Gafaelfawr will no longer send it.  This is not
currently used; right now, TAP looks at the Authorization header,
which is a separate problem since that will no longer be supported
in the future, but that's a problem for another day.