Fix authorization for datalinker
datalinker was using read:image for access control to all of its
non-public routes, but all the routes except /api/datalink/links
are TAP query helpers.  Require read:tap for those and read:image
only for /api/datalink/links.  Remove the conditionals and
values.yaml configuration for scopes and hard-code them, as we do
for other complex configurations.

Also remove the auth-url setting, since we expect datalinker links
to mostly be followed by scripts and the Portal Aspect, for which
redirects to a login page are not useful.