Install helm-docs with go
The homebrew installation method stopped working claiming that brew
is not installed in the Ubuntu image, and setting up homebrew looks
very painful.  Instead, build helm-docs from source with Go and put
the binary in /usr/local/bin.

I tried multiple other ways of downloading the release artifacts
instead of building from source, but they all failed in practice,
getting zero-length files instead of the release artifact that was
promised.  Not sure if this was a temporary GitHub outage or a bug
with the release artifacts of helm-docs.