Write an overview of the repo structure
Since overview/introduction is now an introduction to Phalanx and its
technology ecosystem, overview/repository is now where we can provide an
overview of the repository itself. This content is mostly derived from
the original repository.rst page, but I've moved the details on Helm
charts to the service-guide/service-chart-architecture page so that this
page can remain an overview.

Also stubbed out servic-guide/deploy-from-a-branch, which felt like a
useful topic link from the repository structure page.