Add a Contributing section to overview
There are basic procedures for working with Phalanx that apply to both
service developers and operations specialists that will both work with
the Phalanx repository. Having these procedures in the overview section
could make sense.

An alternative is to put this material in a new top-level section,
Contributing --- but there may not be enough material to justify that.