Integrate Phalanx model into jinja env/template
sphinx-jinja allows us to use Jinja templates to populate content on the
page. In conf.py, we're create a jinja templating environment for every
Phalanx environment. Then on the homepage for an environment, we insert
a templated section using that corresponding Jinja environment for data.
The jinja template is a centrally maintained file,
docs/environments/_summary.rst.jinja. This approach allows us to easily
add the same summary info to each environment's page, customized for
that environment.