Template summary table for applications
Similar to the summary table for environments, this jinja-sphinx-based
template provides summary info for applications in Phalanx.

At the moment the table, and the Application model that powers it, only
gather the app's namespace and cross-links to the environments that use
the application.

Here the code for getting the k8s Namespace is a bit fragile because
it's using a regular expression to parse the namespace from the app's
Helm template. Potentially we could put these namespaces in the
science-platform values to be able to drop use of regular expressions.