Add descriptions to app titles
These descriptions make the /applications/index table of contents more
usable.

I tried customizing the label just in the toctree, but found this also
affects the titles in the sidebar. So in this case, it's better to just
add the descriptions ot the individual pages where they're easier to
maintain.