Add phalanx.lsst.io/docs Chart.yaml annotation
This is a proof-of-concept for adding information about Rubin docs
(technotes and change-controlled documents) to the application summary
tables of apps using Helm's free-form "annotations" field.

Note that items in annotation must be strings. How we handle this is by
making each item a YAML-formatted string that we can parse later.

This is very much based on how Artifact Hub uses annotations
https://artifacthub.io/docs/topics/annotations/helm/