Switch remaining ingresses to GafaelfawrIngress
Switch all the remaining authenticated ingresses except for the
nublado2 configuration, which will continue to have to be
hand-configured to work with Zero-to-JupyterHub, over to using
GafaelfawrIngress to generate the ingress.

In the process, hard-code more things in the GafaelfawrIngress
templates and remove them from values.yaml, since we're
standardizing on not customizing paths, scopes, and similar bits
of the ingress.