Update IP and add tls in values-idfint.yaml
Add TLS auth
Added static broker IPs

Static broker IPs are needed to access the alert-broker from the outside, otherwise every time Kafka is restarted a new is is assigned and the SSL certificates no longer get validated. TLS auth is also needed for the brokers to be able to talk with users.