Switch to new Redis chart for Gafaelfawr
Rather than using the Bitnami Redis chart, switch to the new chart
maintained internally in the charts repository.  This creates a
single Redis instance managed in a stateful set with optional
persistence, equivalent to the Redis code that was already in the
Gafaelfawr chart.

If this works, we'll adopt this chart for all of our other services
that use Redis.