Add autherror location to each NGINX server
The NGINX support for an auth_request subhandler is very limited.
It only allows 401 and 403 returns, and doesn't even include the
WWW-Authenticate header on 403 replies.  We want to use it to do
many more things, so we need to add some custom configuration.

Use server-snippet to add an autherror block to every server in
the NGINX configuration.  This will be used as the target for an
error_page directive introduced by Gafaelfawr and will read the
actual status and error body out of headers.  It will also always
set Cache-Control and WWW-Authenticate.

As is, this block won't be used by anything.  The other half of
this change will introduced Ingress annotations that will use the
block for Ingresses that are Gafaelfawr-protected.