Convert noteburst and times-square to new Redis chart
Use the new internal Redis chart that only creates a single
instance, since in practice we were only using the master created
by the Bitnami chart.  Use SSD storage for Redis on IDF dev.

Although authentication was enabled in values.yaml for noteburst,
Redis authentication was not actually in use so far as I could
determine, so do not configure authentication and rely on the
NetworkPolicy to control access.  (Ideally we should add a secret,
but we can tackle this later after we simplify secret management.)