Use dashes for application names
In the science-platform configuration, use dashes rather than
underscores for the variables that control whether a given
application is installed in a given environment.  This requires
additional quoting in yq and writing the template conditionals
somewhat differently, but makes the naming consistent.

Also use camelCase for more Helm variables in the science-platform
Helm chart.