Fix bootstrapping of ingress-nginx
Work around the fact that nothing sets $auth_status as a variable
until Gafaelfawr runs and creates its first ingress.  The NGINX
documentation seems to indicate that you can set the same variable
multiple times, so ensure that $auth_status always has a value,
which will be overridden by the custom configuration snippets in
the locations that Gafaelfawr is protecting.