Always run docs CI, but filter on path
We want the docs CI run to be blocking for merging so that we don't
merge outright syntax errors, but we only want to do the work if
the documentation changed. Use the same trick as we use with
minikube to always run the workflow but apply path filtering.

Update to Python 3.11 for the documentation build.