Update to Gafaelfawr 9.1.0
Add Helm chart support for quota settings and add quota settings
for the Notebook Aspect to the IDF environments. Switch to camel
case settings so that the Gafaelfawr configuration file looks
closer to the documentation of the Helm chart parameters, and use
with in more places. Reorganize configuration settings to match the
order inside Gafaelfawr for less confusion (although all the
environment settings have not been updated).