Update to latest mobu
Update the mobu configuration for the new 5.0.0 release, which
changes the structure of the flock specifications. Pin IDF dev and
minikube to the testing branch for right now.

Remove the cachemachineImagePolicy setting, since it didn't actually
do anything and cachemachine is going away.

Switch the TAP queries on IDF int back to sync so that we can test
the timeout problems that we are seeing in IDF prod.