Redo how nublado configures JupyterHub
Rather than injecting static files into JupyterHub as secrets,
construct all of our configuration in a ConfigMap and then mount
that into /usr/local/etc/jupyterhub/jupyterhub_config.d. Use that
configuration to set the new parameters for RSPRestSpawner. We now
no longer need a separate Nublado configuration file or all the
parsing code and are uniformly using traitlets.

Mount the Gafaelfawr token for JupyterHub on /etc/gafaelfawr without
a subpath so that hopefully we now get dynamic token updates.

Pin the images for IDF dev to my current patch queue branches.

Fix the indentation for YAML lists in the nublado values.yaml to
be consistent with the rest of Phalanx.