Move remaining nublado configuration
Stop doing any configuration setting through Zero to JupyterHub's
config setting and move the rest of them to the configuration file
we mount in the container. This way we don't have to look for
settings in more than one place and all settings can benefit from
Helm templating.

Import rsp_restspawner to ensure that the objects using traitlets
are registered and can be configured.