Only create nublado-lab-secret if secrets configured
Rather than unconditionally creating the nublado-lab-secret Vault
secret, only create it if we have secrets configured in the lab
configuration. This avoids having to create an empty Vault secret
just to keep vault-secrets-operator happy.

This is not entirely correct since the secret references in the
secrets configuration may not point to this secret, but it will be
good enough for now. We can do fancier Helm templating later if
needed.