Increase JupyterLab startup timeout to 90 seconds
For reasons we don't yet understand, JupyterLab startup sometimes
takes over 60 seconds. Increase the default timeout to 90 seconds
to match what we're currently using in production.

Make the total spawn timeout and the JupyterLab startup timeout
easily configurable in values files. Add a new section to the
values file for hub configuration that we control rather than that
is passed down to Zero to JupyterHub.