Improve nublado configuration documentation
Add helm-docs comments to all of the overrides for Zero to JupyterHub
configuration so that we get more useful chart documentation.

Delete the creation of a new self role. So far as I can tell, this
was never used or assigned to anything, and therefore wasn't doing
anything. I think this was an earlier attempt to solve a problem that
was solved by setting authenticatePrometheus to false.

Change the default permissions of the server role (for spawned labs)
to self rather than inherit. This only affects labs spawned by
JupyterHub administrators and ensures that the lab token only inherits
their regular user permissions and not their admin permissions.