Add Nublado v3 configuration for USDF dev
Add a new configuration for USDF dev modelled after its nublado2
configuration. USDF overrode the ingress annotations, so add support
for setting ingress annotations for the proxy so that it can continue
to do this. That uncovered missed annotations that we added to
nublado2 and not nublado, which have now been copied over.

Rename the ingress to reflect that it's an ingress for the proxy, not
the hub (the proxy routes to either the hub or the lab).