Change TAP schema database pod configuration
Make the TAP schema database pod configuration consistent with other
deployments and Helm values style:

* Use tapSchema rather than tap_schema since Helm uses camelCase
* Add and use podAnnotations, nodeSelector, tolerations, and affinity
  rather than using the ones for the base TAP server
* Add image.pullPolicy configuration

Also set a default image.repository for tapSchema. It would be nice
to have Helm force this to be set, but unfortunately it breaks Renovate
discovery of the latest image tag version, which we want as a reminder
when the image is out of date. Use the same image name that we use for
minikube as the default, but document that it must be set for each
environment.