Increase minimum Python version, update pins
Require at least Python 3.11, matching our other applications.
Update the documenteer version pin for the documentation to not
reference a beta version (which gives pip general permission to
install beta versions) and pin to less than 1.0 for now. Use the
same form of the Pydantic pin that we're using in other
repositories.