Convert to Ruff for linting
Remove the flake8 and isort configuration and their pre-commit
hooks, and add the Ruff pre-commit linter. Add Ruff configuration
based on Gafaelfawr and fix all of the problems that it found in
the src and tests directory. Exclude the installer, since that code
is about to be rewritten.

Enable checks for unresolved merge conflicts in the pre-commit
linter.