Pin dependencies
Since the phalanx command-line tool will now be part of the public
API, we want to pin dependencies and only update them when we know
the new dependencies work. Remove the dependencies from pyproject.toml
and move them to files in requirements/* similar to our other Python
packages. Add new Makefile targets to update the pinned dependencies,
and a make setup target to install only the dependencies required for
the command-line tool.