Add basic Phalanx secret command-line tool
Lay the groundwork for automating Phalanx secret management by adding
a command-line tool. Currently, the only command lists the secrets
that are required for a given environment. Additional commands will
be coming in the future.

This introduces a new set of models separate from the Phalanx models
used by the documentation. These will eventually be merged, but the
models for secret management will require some iteration, so that
merger will be easier once they're in their final form.